Move concrete intersection strategies to plugin layer
