+++ OTAnalytics/application/analysis/intersect.py
@@ -1,67 +1,21 @@
+from abc import ABC, abstractmethod
 from typing import Iterable
 
-from OTAnalytics.application.eventlist import SceneActionDetector, SectionActionDetector
-from OTAnalytics.domain.event import Event, SectionEventBuilder
-from OTAnalytics.domain.intersect import (
-    IntersectAreaByTrackPoints,
-    IntersectBySmallTrackComponents,
-    IntersectImplementation,
-    IntersectParallelizationStrategy,
-)
-from OTAnalytics.domain.section import Area, LineSection, Section
+from OTAnalytics.application.eventlist import SceneActionDetector
+from OTAnalytics.domain.event import Event
+from OTAnalytics.domain.section import Section
 from OTAnalytics.domain.track import Track
 
 
-class RunIntersect:
+class RunIntersect(ABC):
     """
-    This class defines the use case to intersect the given tracks with the given
+    Interface defining the use case to intersect the given tracks with the given
     sections
     """
 
-    def __init__(
-        self,
-        intersect_implementation: IntersectImplementation,
-        intersect_parallelizer: IntersectParallelizationStrategy,
-    ) -> None:
-        self._intersect_implementation = intersect_implementation
-        self._intersect_parallelizer = intersect_parallelizer
-
-    def run(self, tracks: Iterable[Track], sections: Iterable[Section]) -> list[Event]:
-        return self._intersect_parallelizer.execute(
-            self._run_on_single_track, tracks, sections
-        )
-
-    def _run_on_single_track(
-        self,
-        track: Track,
-        sections: Iterable[Section],
-    ) -> list[Event]:
-        events: list[Event] = []
-        for _section in sections:
-            if isinstance(_section, LineSection):
-                line_section_intersector = IntersectBySmallTrackComponents(
-                    implementation=self._intersect_implementation,
-                    line_section=_section,
-                )
-                section_event_builder = SectionEventBuilder()
-                section_action_detector = SectionActionDetector(
-                    intersector=line_section_intersector,
-                    section_event_builder=section_event_builder,
-                )
-            if isinstance(_section, Area):
-                area_section_intersector = IntersectAreaByTrackPoints(
-                    implementation=self._intersect_implementation,
-                    area=_section,
-                )
-                section_event_builder = SectionEventBuilder()
-                section_action_detector = SectionActionDetector(
-                    intersector=area_section_intersector,
-                    section_event_builder=section_event_builder,
-                )
-            events.extend(
-                section_action_detector._detect(section=_section, track=track)
-            )
-        return events
+    @abstractmethod
+    def run(self, track: Iterable[Track], sections: Iterable[Section]) -> list[Event]:
+        raise NotImplementedError
 
 
 class RunSceneEventDetection:
